feat add registration process

+++ src/main/kotlin/de/evoila/bpm/security/repositories/RoleRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.rest.core.annotation.RepositoryRestResource
 
 @RepositoryRestResource(exported = false)
-interface RoleRepository : JpaRepository<UserRole, Int>
+interface RoleRepository : JpaRepository<UserRole, Long>
