+++ src/main/kotlin/de/evoila/bpm/security/repositories/VendorRepository.kt
@@ -3,4 +3,4 @@ package de.evoila.bpm.security.repositories
 import de.evoila.bpm.security.model.Vendor
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface VendorRepository : JpaRepository<Vendor, Int>
+interface VendorRepository : JpaRepository<Vendor, Long>
